#1003 Move init and run commands to neuro_san_studio/commands
