+++ docs/tutorial.md
@@ -100,7 +100,7 @@ Below is a simplified view of the reference project structure. You can adapt it 
 
 * `coded_tools/`: Contains custom-coded tool classes (e.g., `calculator_tool.py`).
 * `logs/`: Where client and server logs are written.
-* `neuro_san_studio/run.py`: A starter script to run the server and the web client.
+* `neuro_san_studio/commands/run.py`: A starter script to run the server and the web client.
 * `registries/`: Holds `.hocon` files that define multi-agent networks and their configurations.
 
 Here are the detailed [instructions](../README.md) to run an agent network along with a web client.
@@ -228,7 +228,7 @@ you also link to or embed the coded tool).
 # Make sure your venv is active
 export AGENT_MANIFEST_FILE="./registries/manifest.hocon"
 export AGENT_TOOL_PATH="./coded_tools"
-python -m neuro_san_studio.run
+python -m neuro_san_studio run
 ```
 
 Since this agent has no further sub-agents or coded tools, it will simply respond to queries but won’t be able to do
@@ -425,7 +425,7 @@ of the LLMs.
 ```bash
 export AGENT_MANIFEST_FILE="./registries/manifest.hocon"
 export AGENT_TOOL_PATH="./coded_tools"
-python -m neuro_san_studio.run
+python -m neuro_san_studio run
 ```
 
 Now, the top-level **Math Geek** agent will parse user queries, pass them to **problem_formulator**, which in turn calls
@@ -948,7 +948,7 @@ of the user guide.
 By default, when you run:
 
 ```bash
-python -m neuro_san_studio.run
+python -m neuro_san_studio run
 ```
 
 * The server logs go to logs/server.log.
@@ -964,7 +964,7 @@ Additionally, you will see logs on your terminal. Checking these files is useful
 
 ## 10. How to Stop the servers
 
-When you are running the server in the foreground (via `python -m neuro_san_studio.run`), simply press:
+When you are running the server in the foreground (via `python -m neuro_san_studio run`), simply press:
 
 * `CTRL + C` on Windows/Mac/Linux terminals
 * This will terminate both the Flask web client server and the `neuro_san` server gracefully.
