+++ docs/user_guide.md
@@ -762,7 +762,7 @@ Simply set the `MODEL_NAME` environment variable before starting the server:
 >
 > ```bash
 > export MODEL_NAME="claude-3-7-sonnet"
-> python -m neuro_san_studio.run
+> python -m neuro_san_studio run
 > ```
 
 ### See also
@@ -1637,7 +1637,7 @@ Furthermore, please install the build requirements in your virtual environment v
     pytest.set_trace()
     ```
 
-3. Start the client and server via `python -m neuro_san_studio.run`,
+3. Start the client and server via `python -m neuro_san_studio run`,
 select `music_nerd_pro` agent network, and ask a question like
 `Where was John Lennon born?`. The code execution stops at the line where you added
 `pytest.set_trace` statement. You can step through the code, view variable values,
