+++ neuro_san_studio/plugins/plugin_loader.py
@@ -17,7 +17,7 @@
 """
 Shared utility for loading plugins from a HOCON configuration file.
 
-Used by both the runner (neuro_san_studio/run.py) and the server wrapper
+Used by both the runner (neuro_san_studio/commands/run.py) and the server wrapper
 (neuro_san_studio/runner/neuro_san_server_wrapper.py) to avoid duplicating the loading logic.
 """
