+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
 ]
 
 [project.scripts]
-neuro-san-studio = "neuro_san_studio.run:main"
+neuro-san-studio = "neuro_san_studio.commands.run:main"
 
 [project.urls]
 Homepage = "https://github.com/cognizant-ai-lab/neuro-san-studio"
